// web/src/gen/types/Collaboration.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Badge } from './Badge.ts'
import type { Opportunity } from './Opportunity.ts'
import type { UserProfile } from './UserProfile.ts'

export type Collaboration = {
  /**
   * @type array | undefined
   */
  badges?: Badge[]
  /**
   * @type string | undefined
   */
  city?: string
  /**
   * @type string | undefined
   */
  country?: string
  /**
   * @type string | undefined
   */
  country_code?: string
  /**
   * @type string | undefined
   */
  created_at?: string
  /**
   * @type string | undefined
   */
  description?: string
  /**
   * @type string | undefined
   */
  hidden_at?: string
  /**
   * @type integer | undefined
   */
  id?: number
  /**
   * @type boolean | undefined
   */
  is_payable?: boolean
  /**
   * @type object | undefined
   */
  opportunity?: Opportunity
  /**
   * @type integer | undefined
   */
  opportunity_id?: number
  /**
   * @type string | undefined
   */
  published_at?: string
  /**
   * @type string | undefined
   */
  title?: string
  /**
   * @type object | undefined
   */
  user?: UserProfile
  /**
   * @type integer | undefined
   */
  user_id?: number
}